Adding VIMRC
This is the VIMRC I have been using pretty regularly now.  It's in a
constant state of motion but does seem to be working.  Parts of it have
been pulled from various locations on the internet, coworkers, and a
few I've just learned over the years. Some recent behavior changes have
been heavily influenced by the SPF13 project in how to better handle
plugins and friends.

Signed-off-by: Dan Kalowsky <dank@deadmime.org>