vimrc: add noinsert/noselect to completeopt
Make vim work more like a traditional IDE where it doesn't auto insert
the value that you don't want.  To do this we make sure the completeopt
value has the noinsert and noselect value enabled.  As we tune this, it
may be better to have the select enabled and disable the insert only

Signed-off-by: Dan Kalowsky <dank@deadmime.org>