gitconfig: correct alias for interactive commits
The previous alias, icommit, didn't work correctly with the githooks.
It would still allow the clang_format to run over the whole of the file
and then stash those changes, hiding them from view.

Changed the name from icommit, to pcommit, to better reflect that this
is a patched commit.  Now correctly git stash and saving the index and
then popping it off at the end of the commit.

Signed-off-by: Dan Kalowsky <dank@deadmime.org>