vimrc: keep default clipboard when ssh in
When ssh'ing in, the default unnamed clipboard becomes * not, " which
causes problems when trying to yank, delete, copy, and paste.

The previous setup could still work as long as you'd paste with
something like ""p  (to pull from the correct register).  Instead of
having to remember that, check if the SSH_CLIENT environment variable
is set, and if it is, do not allow the resetting of the clipboard value.

Signed-off-by: Dan Kalowsky <dank@deadmime.org>