vimrc: use plugs to check if loaded versus file
The Plug extension has a mechanism to check what plugins have been
loaded.  Using this works better than checking if a file has been
downloaded.

I've had a few cases I've disabled a plugin in the vimrc.bundles.local
file but forgotten to remove the checkout in .vim.  As such the
extension continues to load.

Signed-off-by: Dan Kalowsky <dank@deadmime.org>