vimrc: move from airline to lightline
airline is really nice, but I'm getting a little annoyed at the hacked
up font problem.  More specifically that hacked up font problem on MSYS
and Windows Subsystem Linux.  Moving to lightline as it can provide a
lot of the same functionality, but won't need to look as pretty with a
hacked up font.  So far, I've been just as happy with it.

Signed-off-by: Dan Kalowsky <dank@deadmime.org>