bootstrap: remove the installStow function
Previously I'd written a function to install stow when it wasn't
available.  This turns out to be a bad idea if the host system doesn't
actually have any command line compile tools installed.

Instead of depending upon compiling a tool, just use the version written
for operations on Windows and move along.

Signed-off-by: Dan Kalowsky <dank@deadmime.org>