bootstrap: change uname to switch statement
The original method worked fine on Linux and macOS, but when you add in
MSYS or Cygwin, things go south.  The obvious answer is to use the uname
-o flag, but macOS doesn't support that, so the next best possible
solution is to break it out by switch statement.

Future work my consolidate all the Windows pieces to a single platform
and then add a new variable for further detail (PLAT_IMPL?).

Signed-off-by: Dan Kalowsky <dkalowsky@amperecomputing.com>