bashrc: separate interactive and non-interactive commands
There is a need to have these commands separated.  It is exacerbated by
the fact that SSH has the ForwardAgent command for ssh, but not for scp
(pre version 8.4).

Use the check for PS1 as a mechanism to determine if an interactive
shell is being started up.  If it is, we want to source the new
.bashrc-interactive file, otherwise we will not do anything.

Signed-off-by: Dan Kalowsky <dank@deadmime.org>