X11: add override for gameball
The gameball trackball is... unique.  So far I'm not liking the button
placement.  Use the X11 setup to override the values and configure
the buttons the way I want... where the touch on the left is vertical
and right is horizontal, while the thumb cluster is the primary click
and right click leaving the ring finger as the forward and back.

Signed-off-by: Dan Kalowsky <dank@deadmime.org>