nvim: add options.lua
Add a file to contain basic configuration options for nvim
which is largely pulled from my previous vimrc.

New addition, keep the list in alphabetical order to make it
easier to find things.  Not sure if that is going to be any
help or not, but we'll give it a go.

Signed-off-by: Dan Kalowsky <dank@deadmime.org>