nvim: add guard for formating
I thought I had added in some tooling to run formatting, but
it seems to be not there.

Enable guard.lua to do this.  Downfall here is you have to
hard code the enabling of formatters and linters which is
standard but it seems there is no way to check if a project
has a .clang-format file and if so enable that linter only
then.

Signed-off-by: Dan Kalowsky <dank@deadmime.org>