bashrc: move bash_profile to dan_profile
Using a new terminal emulator is showing some issues with how I've got
the .bash_profile setup.  So... just have the .bash_profile parse the
.bashrc Now have the .bashrc parse the newly created .dan_profile.
Additionally make sure to only load this when we are in an interactive
shell.

Additional fix in the .dan_profile is to use tput to check the color
support on the terminal rather than actually needing to know the
terminal type.

Signed-off-by: Dan Kalowsky <dank@deadmime.org>