bootstrap: convert doStow to find loop
Instead of trying to manually do the loop, just use
find for the all the files in provided directory to go
map to the home dir.

This has two big impacts.

The first is the code gets a lot simpler to do as
special files are no longer being included here.

The second is that it changes the basic behavior to
keep a user's home with the directories it needs, it
just overrides the specific files with the ones I want
to use.  This means for example, .config will not be
linked to my .config anymore, but .config/nvim/lua/init.lua
will be linked to my config/.config/nvim/lua/init.lua.

This limits the pollution from new processes adding to
.dirs.

Signed-off-by: Dan Kalowsky <dkalowsky@amperecomputing.com>