nvim: osc52: reduce usage
While all of this works really well for remot SSH copy,
on a local system it breaks copying to the clipboard
and instead copies the whole file... ocassionally.

I did something bad, but I can't figure out what it is
so for now revert some of it.

Hopefully this can all go away when updating to nvim v10
where osc52 is native to the system.

Signed-off-by: Dan Kalowsky <dank@deadmime.org>